@@ -1,17 +1,20 @@
 // this barrel export is currently only used to find undocumented exports
+export * from './packages/codemirror/index.mjs';
 export * from './packages/core/index.mjs';
 export * from './packages/csound/index.mjs';
+export * from './packages/desktopbridge/index.mjs';
+export * from './packages/draw/index.mjs';
 export * from './packages/embed/index.mjs';
-export * from './packages/eval/index.mjs';
+export * from './packages/hydra/index.mjs';
 export * from './packages/midi/index.mjs';
 export * from './packages/mini/index.mjs';
 export * from './packages/osc/index.mjs';
-export * from './packages/react/index.mjs';
+export * from './packages/repl/index.mjs';
 export * from './packages/serial/index.mjs';
 export * from './packages/soundfonts/index.mjs';
+export * from './packages/superdough/index.mjs';
 export * from './packages/tonal/index.mjs';
-export * from './packages/tone/index.mjs';
 export * from './packages/transpiler/index.mjs';
+export * from './packages/web/index.mjs';
 export * from './packages/webaudio/index.mjs';
-export * from './packages/webdirt/index.mjs';
 export * from './packages/xen/index.mjs';
