@@ -7,6 +7,9 @@ function createClock(
   duration = 0.05, // duration of each cycle
   interval = 0.1, // interval between callbacks
   overlap = 0.1, // overlap between callbacks
+  setInterval = globalThis.setInterval,
+  clearInterval = globalThis.clearInterval,
+  round = true,
 ) {
   let tick = 0; // counts callbacks
   let phase = 0; // next callback time
@@ -22,9 +25,8 @@ function createClock(
     }
     // callback as long as we're inside the lookahead
     while (phase < lookahead) {
-      phase = Math.round(phase * precision) / precision;
-      phase >= t && callback(phase, duration, tick);
-      phase < t && console.log('TOO LATE', phase); // what if latency is added from outside?
+      phase = round ? Math.round(phase * precision) / precision : phase;
+      callback(phase, duration, tick, t); // callback has to skip / handle phase < t!
       phase += duration; // increment phase by duration
       tick++;
     }
@@ -35,7 +37,10 @@ function createClock(
     onTick();
     intervalID = setInterval(onTick, interval * 1000);
   };
-  const clear = () => intervalID !== undefined && clearInterval(intervalID);
+  const clear = () => {
+    intervalID !== undefined && clearInterval(intervalID);
+    intervalID = undefined;
+  };
   const pause = () => clear();
   const stop = () => {
     tick = 0;
