@@ -4,9 +4,9 @@ class Strudel extends HTMLElement {
   }
   connectedCallback() {
     setTimeout(() => {
-      const code = (this.innerHTML + '').replace('<!--', '').replace('-->', '').trim();
+      const code = this.getAttribute('code') || (this.innerHTML + '').replace('<!--', '').replace('-->', '').trim();
       const iframe = document.createElement('iframe');
-      const src = `https://strudel.tidalcycles.org/#${encodeURIComponent(btoa(code))}`;
+      const src = `https://strudel.cc/#${encodeURIComponent(btoa(code))}`;
       // const src = `http://localhost:3000/#${encodeURIComponent(btoa(code))}`;
       iframe.setAttribute('src', src);
       iframe.setAttribute('width', '600');
