@@ -8,8 +8,8 @@ export default defineConfig({
   build: {
     lib: {
       entry: resolve(__dirname, 'osc.mjs'),
-      formats: ['es', 'cjs'],
-      fileName: (ext) => ({ es: 'index.mjs', cjs: 'index.js' }[ext]),
+      formats: ['es'],
+      fileName: (ext) => ({ es: 'index.mjs' })[ext],
     },
     rollupOptions: {
       external: [...Object.keys(dependencies)],
