@@ -5,7 +5,27 @@ This program is free software: you can redistribute it and/or modify it under th
 */
 
 import { Note, Interval, Scale } from '@tonaljs/tonal';
-import { register, _mod } from '@strudel.cycles/core';
+import { register, _mod, silence, logger, pure, isNote } from '@strudel/core';
+import { stepInNamedScale } from './tonleiter.mjs';
+
+const octavesInterval = (octaves) => (octaves <= 0 ? -1 : 1) + octaves * 7 + 'P';
+
+function scaleStep(step, scale) {
+  scale = scale.replaceAll(':', ' ');
+  step = Math.ceil(step);
+  let { intervals, tonic, empty } = Scale.get(scale);
+  if ((empty && isNote(scale)) || (!empty && !tonic)) {
+    throw new Error(`incomplete scale. Make sure to use ":" instead of spaces, example: .scale("C:major")`);
+  } else if (empty) {
+    throw new Error(`invalid scale "${scale}"`);
+  }
+  tonic = tonic || 'C';
+  const { pc, oct = 3 } = Note.get(tonic);
+  const octaveOffset = Math.floor(step / intervals.length);
+  const scaleStep = _mod(step, intervals.length);
+  const interval = Interval.add(intervals[scaleStep], octavesInterval(octaveOffset));
+  return Note.transpose(pc + oct, interval);
+}
 
 // transpose note inside scale by offset steps
 // function scaleOffset(scale: string, offset: number, note: string) {
@@ -76,16 +96,37 @@ function scaleOffset(scale, offset, note) {
 
 export const transpose = register('transpose', function (intervalOrSemitones, pat) {
   return pat.withHap((hap) => {
-    const interval = !isNaN(Number(intervalOrSemitones))
-      ? Interval.fromSemitones(intervalOrSemitones /*  as number */)
-      : String(intervalOrSemitones);
-    if (typeof hap.value === 'number') {
-      const semitones = typeof interval === 'string' ? Interval.semitones(interval) || 0 : interval;
-      return hap.withValue(() => hap.value + semitones);
+    const note = hap.value.note ?? hap.value;
+    if (typeof note === 'number') {
+      // note is a number, so just add the number semitones of the interval
+      let semitones;
+      if (typeof intervalOrSemitones === 'number') {
+        semitones = intervalOrSemitones;
+      } else if (typeof intervalOrSemitones === 'string') {
+        semitones = Interval.semitones(intervalOrSemitones) || 0;
+      }
+      const targetNote = note + semitones;
+      if (typeof hap.value === 'object') {
+        return hap.withValue(() => ({ ...hap.value, note: targetNote }));
+      }
+      return hap.withValue(() => targetNote);
     }
+    if (typeof note !== 'string' || !isNote(note)) {
+      logger(`[tonal] transpose: not a note "${note}"`, 'warning');
+      return hap;
+    }
+    // note is a string, so we might be able to preserve harmonics if interval is a string as well
+    const interval = !isNaN(Number(intervalOrSemitones))
+      ? Interval.fromSemitones(intervalOrSemitones)
+      : String(intervalOrSemitones);
     // TODO: move simplify to player to preserve enharmonics
     // tone.js doesn't understand multiple sharps flats e.g. F##3 has to be turned into G3
-    return hap.withValue(() => Note.simplify(Note.transpose(hap.value, interval)));
+    // TODO: check if this is still relevant..
+    const targetNote = Note.simplify(Note.transpose(note, interval));
+    if (typeof hap.value === 'object') {
+      return hap.withValue(() => ({ ...hap.value, note: targetNote }));
+    }
+    return hap.withValue(() => targetNote);
   });
 });
 
@@ -113,6 +154,11 @@ export const scaleTranspose = register('scaleTranspose', function (offset /* : n
     if (!hap.context.scale) {
       throw new Error('can only use scaleTranspose after .scale');
     }
+    if (typeof hap.value === 'object')
+      return hap.withValue(() => ({
+        ...hap.value,
+        note: scaleOffset(hap.context.scale, Number(offset), hap.value.note),
+      }));
     if (typeof hap.value !== 'string') {
       throw new Error('can only use scaleTranspose with notes');
     }
@@ -127,37 +173,81 @@ export const scaleTranspose = register('scaleTranspose', function (offset /* : n
  *
  * The root note defaults to octave 3, if no octave number is given.
  *
- * @memberof Pattern
  * @name scale
  * @param {string} scale Name of scale
  * @returns Pattern
  * @example
- * "0 2 4 6 4 2".scale("C2:major").note()
+ * n("0 2 4 6 4 2").scale("C:major")
  * @example
- * "0 2 4 6 4 2"
- * .scale("C2:<major minor>")
- * .note()
+ * n("[0,7] 4 [2,7] 4")
+ * .scale("C:<major minor>/2")
+ * .s("piano")
  * @example
- * "0 1 2 3 4 5 6 7".rev().scale("C2:<major minor>").note()
- * .s("folkharp")
+ * n(rand.range(0,12).segment(8))
+ * .scale("C:ritusen")
+ * .s("piano")
  */
 
-export const scale = register('scale', function (scale, pat) {
-  // Supports ':' list syntax in mininotation
-  if (Array.isArray(scale)) {
-    scale = scale.join(' ');
-  }
-  return pat.withHap((hap) => {
-    const isObject = typeof hap.value === 'object';
-    let note = isObject ? hap.value.n : hap.value;
-    const asNumber = Number(note);
-    if (!isNaN(asNumber)) {
-      // TODO: worth keeping for supporting ':' in (non-mininotation) strings?
-      scale = scale.replaceAll(':', ' ');
-      let [tonic, scaleName] = Scale.tokenize(scale);
-      const { pc, oct = 3 } = Note.get(tonic);
-      note = scaleOffset(pc + ' ' + scaleName, asNumber, pc + oct);
+export const scale = register(
+  'scale',
+  function (scale, pat) {
+    // Supports ':' list syntax in mininotation
+    if (Array.isArray(scale)) {
+      scale = scale.flat().join(' ');
     }
-    return hap.withValue(() => (isObject ? { ...hap.value, note } : note)).setContext({ ...hap.context, scale });
-  });
-});
+    return (
+      pat
+        .fmap((value) => {
+          const isObject = typeof value === 'object';
+          let step = isObject ? value.n : value;
+          if (isObject) {
+            delete value.n; // remove n so it won't cause trouble
+          }
+          if (isNote(step)) {
+            // legacy..
+            return pure(step);
+          }
+          let asNumber = Number(step);
+          let semitones = 0;
+          if (isNaN(asNumber)) {
+            step = String(step);
+            if (!/^[-+]?\d+(#*|b*){1}$/.test(step)) {
+              logger(
+                `[tonal] invalid scale step "${step}", expected number or integer with optional # b suffixes`,
+                'error',
+              );
+              return silence;
+            }
+            const isharp = step.indexOf('#');
+            if (isharp >= 0) {
+              asNumber = Number(step.substring(0, isharp));
+              semitones = step.length - isharp;
+            } else {
+              const iflat = step.indexOf('b');
+              asNumber = Number(step.substring(0, iflat));
+              semitones = iflat - step.length;
+            }
+          }
+          try {
+            let note;
+            if (isObject && value.anchor) {
+              note = stepInNamedScale(asNumber, scale, value.anchor);
+            } else {
+              note = scaleStep(asNumber, scale);
+            }
+            if (semitones != 0) note = Note.transpose(note, Interval.fromSemitones(semitones));
+            value = pure(isObject ? { ...value, note } : note);
+          } catch (err) {
+            logger(`[tonal] ${err.message}`, 'error');
+            value = silence;
+          }
+          return value;
+        })
+        .outerJoin()
+        // legacy:
+        .withHap((hap) => hap.setContext({ ...hap.context, scale }))
+    );
+  },
+  true,
+  true, // preserve tactus
+);
