@@ -11,22 +11,20 @@ const simple = { wrapAsync: false, addReturn: false, simpleLocs: true };
 
 describe('transpiler', () => {
   it('wraps double quote string with mini and adds location', () => {
-    expect(transpiler('"c3"', simple)).toEqual("mini('c3').withMiniLocation(0, 4);");
-    expect(transpiler('stack("c3","bd sd")', simple)).toEqual(
-      "stack(mini('c3').withMiniLocation(6, 10), mini('bd sd').withMiniLocation(11, 18));",
-    );
+    expect(transpiler('"c3"', simple).output).toEqual("m('c3', 0);");
+    expect(transpiler('stack("c3","bd sd")', simple).output).toEqual("stack(m('c3', 6), m('bd sd', 11));");
   });
   it('wraps backtick string with mini and adds location', () => {
-    expect(transpiler('`c3`', simple)).toEqual("mini('c3').withMiniLocation(0, 4);");
-  });
-  it('replaces note variables with note strings', () => {
-    expect(transpiler('seq(c3, d3)', simple)).toEqual("seq('c3', 'd3');");
+    expect(transpiler('`c3`', simple).output).toEqual("m('c3', 0);");
   });
   it('keeps tagged template literal as is', () => {
-    expect(transpiler('xxx`c3`', simple)).toEqual('xxx`c3`;');
+    expect(transpiler('xxx`c3`', simple).output).toEqual('xxx`c3`;');
   });
   it('supports top level await', () => {
-    expect(transpiler("await samples('xxx');", simple)).toEqual("await samples('xxx');");
+    expect(transpiler("await samples('xxx');", simple).output).toEqual("await samples('xxx');");
+  });
+  it('adds await to bare samples call', () => {
+    expect(transpiler("samples('xxx');", simple).output).toEqual("await samples('xxx');");
   });
   /*   it('parses dynamic imports', () => {
     expect(
@@ -36,4 +34,12 @@ describe('transpiler', () => {
       }),
     ).toEqual("const {default: foo} = await import('https://bar.com/foo.js');");
   }); */
+  it('collections locations', () => {
+    const { miniLocations } = transpiler(`s("bd", "hh oh")`, { ...simple, emitMiniLocations: true });
+    expect(miniLocations).toEqual([
+      [3, 5],
+      [9, 11],
+      [12, 14],
+    ]);
+  });
 });
