@@ -3,33 +3,24 @@
 // it might require mocking more stuff when tunes added that use other functions
 
 // import * as tunes from './tunes.mjs';
-// import { evaluate } from '@strudel.cycles/eval';
-import { evaluate } from '@strudel.cycles/transpiler';
-import { evalScope } from '@strudel.cycles/core';
-import * as strudel from '@strudel.cycles/core';
-import * as webaudio from '@strudel.cycles/webaudio';
-import controls from '@strudel.cycles/core/controls.mjs';
-// import gist from '@strudel.cycles/core/gist.js';
-import { mini } from '@strudel.cycles/mini/mini.mjs';
-// import * as toneHelpers from '@strudel.cycles/tone/tone.mjs';
-// import * as voicingHelpers from '@strudel.cycles/tonal/voicings.mjs';
-// import * as uiHelpers from '@strudel.cycles/tone/ui.mjs';
-// import * as drawHelpers from '@strudel.cycles/tone/draw.mjs';
-// import euclid from '@strudel.cycles/core/euclid.mjs';
-// import '@strudel.cycles/tone/tone.mjs';
-// import '@strudel.cycles/midi/midi.mjs';
-import * as tonalHelpers from '@strudel.cycles/tonal';
-import '@strudel.cycles/xen/xen.mjs';
-// import '@strudel.cycles/xen/tune.mjs';
-// import '@strudel.cycles/core/euclid.mjs';
-// import '@strudel.cycles/core/speak.mjs'; // window is not defined
-// import '@strudel.cycles/tone/pianoroll.mjs';
-// import '@strudel.cycles/tone/draw.mjs';
-// import '@strudel.cycles/osc/osc.mjs';
-// import '@strudel.cycles/webaudio/webaudio.mjs';
-// import '@strudel.cycles/serial/serial.mjs';
-// import controls from '@strudel.cycles/core/controls.mjs';
-import '../website/src/repl/prebake';
+import { evaluate } from '@strudel/transpiler';
+import { evalScope } from '@strudel/core';
+import * as strudel from '@strudel/core';
+import * as webaudio from '@strudel/webaudio';
+// import gist from '@strudel/core/gist.js';
+import { mini, m } from '@strudel/mini/mini.mjs';
+// import * as voicingHelpers from '@strudel/tonal/voicings.mjs';
+// import euclid from '@strudel/core/euclid.mjs';
+// import '@strudel/midi/midi.mjs';
+import * as tonalHelpers from '@strudel/tonal';
+import '@strudel/xen/xen.mjs';
+// import '@strudel/xen/tune.mjs';
+// import '@strudel/core/euclid.mjs';
+// import '@strudel/core/speak.mjs'; // window is not defined
+// import '@strudel/osc/osc.mjs';
+// import '@strudel/webaudio/webaudio.mjs';
+// import '@strudel/serial/serial.mjs';
+import '../website/src/repl/piano';
 
 class MockedNode {
   chain() {
@@ -67,6 +58,7 @@ const toneHelpersMocked = {
   vol: mockNode,
   out: id,
   osc: id,
+  samples: id,
   adsr: id,
   getDestination: id,
   players: mockNode,
@@ -130,6 +122,19 @@ strudel.Pattern.prototype.midi = function () {
   return this;
 };
 
+strudel.Pattern.prototype._scope = function () {
+  return this;
+};
+strudel.Pattern.prototype._spiral = function () {
+  return this;
+};
+strudel.Pattern.prototype._pitchwheel = function () {
+  return this;
+};
+strudel.Pattern.prototype._pianoroll = function () {
+  return this;
+};
+
 const uiHelpersMocked = {
   backgroundImage: id,
 };
@@ -159,10 +164,9 @@ evalScope(
   strudel,
   toneHelpersMocked,
   uiHelpersMocked,
-  controls,
   webaudio,
   tonalHelpers,
-  /* controls,
+  /*
   toneHelpers,
   voicingHelpers,
   drawHelpers,
@@ -174,20 +178,23 @@ evalScope(
     csound: id,
     loadOrc: id,
     mini,
+    m,
     getDrawContext,
     getAudioContext,
     loadSoundfont,
     loadCSound,
     loadCsound,
     loadcsound,
+    setcps: id,
     Clock: {}, // whatever
     // Tone,
   },
 );
 
+// TBD: use transpiler to support labeled statements
 export const queryCode = async (code, cycles = 1) => {
   const { pattern } = await evaluate(code);
-  const haps = pattern.queryArc(0, cycles);
+  const haps = pattern.sortHapsByPart().queryArc(0, cycles);
   return haps.map((h) => h.show(true));
 };
 
@@ -221,7 +228,7 @@ export const testCycles = {
   festivalOfFingers: 16,
   festivalOfFingers2: 22,
   undergroundPlumber: 20,
-  bridgeIsOver: 16,
+  bridgeIsOver: 8,
   goodTimes: 16,
   echoPiano: 8,
   sml1: 48,
@@ -229,9 +236,8 @@ export const testCycles = {
   randomBells: 24,
   waa: 16,
   waar: 16,
-  hyperpop: 10,
   festivalOfFingers3: 16,
 };
 
-// fixed: https://strudel.tidalcycles.org/?DBp75NUfSxIn (missing .note())
-// bug: https://strudel.tidalcycles.org/?xHaKTd1kTpCn + https://strudel.tidalcycles.org/?o5LLePbx8kiQ
+// fixed: https://strudel.cc/?DBp75NUfSxIn (missing .note())
+// bug: https://strudel.cc/?xHaKTd1kTpCn + https://strudel.cc/?o5LLePbx8kiQ
