@@ -2,28 +2,25 @@ import jsdoc from '../../../doc.json'; // doc.json is built with `npm run jsdoc-
 const docs = jsdoc.docs.reduce((acc, obj) => Object.assign(acc, { [obj.longname]: obj }), {});
 import { MiniRepl } from './MiniRepl';
 
-const getTag = (title, item) => item.tags?.find((t) => t.title === title)?.text;
-
 export function JsDoc({ name, h = 3, hideDescription, punchcard, canvasHeight }) {
   const item = docs[name];
   if (!item) {
     console.warn('Not found: ' + name);
     return <div />;
   }
-  const synonyms = getTag('synonyms', item)?.split(', ') || [];
   const CustomHeading = `h${h}`;
-  const description = item.description.replaceAll(/\{@link ([a-zA-Z\.]+)?#?([a-zA-Z]*)\}/g, (_, a, b) => {
-    // console.log(_, 'a', a, 'b', b);
-    return `<a href="#${a.replaceAll('.', '').toLowerCase()}${b ? `-${b}` : ''}">${a}${b ? `#${b}` : ''}</a>`;
-  });
+  const description =
+    item.description?.replaceAll(/\{@link ([a-zA-Z\.]+)?#?([a-zA-Z]*)\}/g, (_, a, b) => {
+      return `<a href="#${a.replaceAll('.', '').toLowerCase()}${b ? `-${b}` : ''}">${a}${b ? `#${b}` : ''}</a>`;
+    }) || '';
   return (
     <>
       {!!h && <CustomHeading>{item.longname}</CustomHeading>}
       {!hideDescription && (
         <>
-          {!!synonyms.length && (
+          {!!item.synonyms_text && (
             <span>
-              Synonyms: <code>{synonyms.join(', ')}</code>
+              Synonyms: <code>{item.synonyms_text}</code>
             </span>
           )}
           <div dangerouslySetInnerHTML={{ __html: description }} />
